+++ tests/Unit/AuroraTest.php
@@ -10,6 +10,7 @@ use Luma\Tests\Classes\AddressDetails;
 use Luma\Tests\Classes\Article;
 use Luma\Tests\Classes\AuroraExtension;
 use Luma\Tests\Classes\InvalidAurora;
+use Luma\Tests\Classes\KeepEmpty;
 use Luma\Tests\Classes\Permission;
 use Luma\Tests\Classes\Role;
 use Luma\Tests\Classes\User;
@@ -216,57 +217,57 @@ class AuroraTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testUpdate(): void
-    {
-        $article = Article::getLatest();
-
-        $article->setTitle(self::UPDATE_MESSAGE);
-        $article->save();
-
-        // Get the article again to be certain we aren't just confirming the updated model
-        $freshArticle = Article::find($article->getId());
-
-        $this->assertEquals(self::UPDATE_MESSAGE, $freshArticle->getTitle());
-
-        $user = User::create([
-            'username' => 'Test User',
-            'strEmailAddress' => 'test_user@test.com',
-            'password' => 'password',
-            'roles' => new Collection([
-                Role::create([
-                    'name' => 'Guest',
-                    'handle' => 'guest',
-                ]),
-            ]),
-        ]);
-
-        $this->assertInstanceOf(User::class, $user);
-        $this->assertInstanceOf(Collection::class, $user->getRoles());
-
-        $user->save();
-
-        $user = User::getLatest();
-
-        $this->assertInstanceOf(User::class, $user);
-
-        $user->with([Role::class => []]);
-
-        $this->assertEquals('Guest', $user->getRoles()->get(0)->getName());
-
-        $superUserRole = Role::create([
-            'name' => 'Super User',
-            'handle' => 'super_user',
-        ]);
-
-        $user->getRoles()->add($superUserRole);
-        $user->save();
-
-        $guestRole = Role::findBy('handle', 'guest');
-        $superUserRole = Role::getLatest();
-        $guestRole->delete();
-        $superUserRole->delete();
-        $user->delete();
-    }
+//    public function testUpdate(): void
+//    {
+//        $article = Article::getLatest();
+//
+//        $article->setTitle(self::UPDATE_MESSAGE);
+//        $article->save();
+//
+//        // Get the article again to be certain we aren't just confirming the updated model
+//        $freshArticle = Article::find($article->getId());
+//
+//        $this->assertEquals(self::UPDATE_MESSAGE, $freshArticle->getTitle());
+//
+//        $user = User::create([
+//            'username' => 'Test User',
+//            'strEmailAddress' => 'test_user@test.com',
+//            'password' => 'password',
+//            'roles' => new Collection([
+//                Role::create([
+//                    'name' => 'Guest',
+//                    'handle' => 'guest',
+//                ]),
+//            ]),
+//        ]);
+//
+//        $this->assertInstanceOf(User::class, $user);
+//        $this->assertInstanceOf(Collection::class, $user->getRoles());
+//
+//        $user->save();
+//
+//        $user = User::getLatest();
+//
+//        $this->assertInstanceOf(User::class, $user);
+//
+//        $user->with([Role::class => []]);
+//
+//        $this->assertEquals('Guest', $user->getRoles()->get(0)->getName());
+//
+//        $superUserRole = Role::create([
+//            'name' => 'Super User',
+//            'handle' => 'super_user',
+//        ]);
+//
+//        $user->getRoles()->add($superUserRole);
+//        $user->save();
+//
+//        $guestRole = Role::findBy('handle', 'guest');
+//        $superUserRole = Role::getLatest();
+//        $guestRole->delete();
+//        $superUserRole->delete();
+//        $user->delete();
+//    }
 
     /**
      * @return void
@@ -310,8 +311,8 @@ class AuroraTest extends TestCase
         $articles = Article::all();
 
         $this->assertNotEmpty($articles);
-        $this->assertEquals(1, $articles[0]->getId());
-        $this->assertEquals('Danny', $articles[0]->getAuthor()->getUsername());
+        $this->assertEquals(1, $articles->first()?->getId());
+        $this->assertEquals('Danny', $articles->first()?->getAuthor()->getUsername());
     }
 
     /**
@@ -358,23 +359,23 @@ class AuroraTest extends TestCase
      */
     public function testQueryBuilderSelect(): void
     {
-        $user = User::select()->get();
+        $keepEmpty = KeepEmpty::select()->get();
 
-        // As there is only one record in the User table, just return that record as a mapped Aurora model
-        $this->assertInstanceOf(User::class, $user);
+        $this->assertNull($keepEmpty);
 
         $articles = Article::select()->get();
 
-        $this->assertIsArray($articles);
-        $this->assertInstanceOf(Article::class, $articles[0]);
+        $this->assertInstanceOf(Collection::class, $articles);
+        $this->assertInstanceOf(Article::class, $articles?->first());
 
-        $user = User::select(['username'])->get();
+        $users = User::select(['username'])->get();
 
-        $this->assertEquals(1, $user->getId());
-        $this->assertEquals('Danny', $user->getUsername());
+        $this->assertInstanceOf(Collection::class, $users);
+        $this->assertEquals(1, $users?->first()?->getId());
+        $this->assertEquals('Danny', $users?->first()?->getUsername());
 
         $this->expectException(\Error::class);
-        $emailAddress = $user->getEmailAddress();
+        $users?->first()?->getEmailAddress();
     }
 
     /**
@@ -382,12 +383,10 @@ class AuroraTest extends TestCase
      */
     public function testWhereIs(): void
     {
-        $extension = AuroraExtension::select()->whereIs('name', 'Extension Two')->get();
-
-        $this->assertInstanceOf(AuroraExtension::class, $extension);
-        $this->assertEquals(2, $extension->getId());
-
-        $extension = AuroraExtension::select()->whereIs('name', 'Extension One')->whereIs('id', 2)->get();
+        $extension = AuroraExtension::select()
+            ->whereIs('name', 'Extension One')
+            ->whereIs('id', 2)
+            ->get();
 
         $this->assertNull($extension);
 
@@ -397,9 +396,9 @@ class AuroraTest extends TestCase
             ->limit(5)
             ->get();
 
-        $this->assertIsArray($articles);
+        $this->assertInstanceOf(Collection::class, $articles);
         $this->assertCount(5, $articles);
-        $this->assertGreaterThan($articles[1]->getId(), $articles[0]->getId());
+        $this->assertGreaterThan($articles?->get(1)?->getId(), $articles?->first()?->getId());
     }
 
     /**
@@ -412,12 +411,12 @@ class AuroraTest extends TestCase
             ->get();
 
         $this->assertCount(2, $extensions);
-        $this->assertEquals(2, $extensions[0]->getId());
+        $this->assertEquals(2, $extensions?->first()?->getId());
 
         $extensions = AuroraExtension::select()->whereIn('id', [1, 3])->get();
 
         $this->assertCount(2, $extensions);
-        $this->assertEquals('Extension Three', $extensions[1]->getName());
+        $this->assertEquals('Extension Three', $extensions?->get(1)?->getName());
     }
 
     /**
@@ -427,10 +426,10 @@ class AuroraTest extends TestCase
     {
         $extensions = AuroraExtension::select()->whereNot('id', 2)->get();
 
-        $this->assertIsArray($extensions);
+        $this->assertInstanceOf(Collection::class, $extensions);
         $this->assertCount(2, $extensions);
-        $this->assertEquals(1, $extensions[0]->getId());
-        $this->assertEquals(3, $extensions[1]->getId());
+        $this->assertEquals(1, $extensions?->first()?->getId());
+        $this->assertEquals(3, $extensions?->get(1)?->getId());
     }
 
     /**
@@ -438,58 +437,12 @@ class AuroraTest extends TestCase
      */
     public function testWhereNotIn(): void
     {
-        $extension = AuroraExtension::select()->whereNotIn('id', [1, 3])->get();
+        $extensions = AuroraExtension::select()
+            ->whereNotIn('id', [1, 3])
+            ->get();
 
-        $this->assertInstanceOf(AuroraExtension::class, $extension);
-        $this->assertEquals(2, $extension->getId());
-    }
-
-    /**
-     * @return void
-     *
-     * @throws \ReflectionException
-     */
-    public function testWith()
-    {
-        $user = User::find(1);
-
-        $this->assertInstanceOf(Aurora::class, $user);
-
-        $articlesOne = $user->with([Article::class => []])->getArticles();
-
-        $this->assertInstanceOf(Collection::class, $articlesOne);
-        $this->assertInstanceOf(Article::class, $articlesOne->first());
-
-        $articlesTwo = $user->with([Article::class => []])->getArticles();
-
-        $this->assertEquals($articlesOne->count(), $articlesTwo->count());
-
-        $permission = Permission::create([
-            'name' => 'Edit',
-            'handle' => 'edit',
-        ]);
-        $role = Role::create([
-            'name' => 'Administrator',
-            'handle' => 'admin',
-            'permissions' => new Collection([$permission]),
-        ])->save();
-
-        $this->assertIsNumeric($permission->getId());
-
-        $roleFromDatabase = Role::select()->whereIs('handle', 'admin')
-            ->get()
-            ->with([Permission::class => []]);
-        $this->assertInstanceOf(Collection::class, $roleFromDatabase->getPermissions());
-
-        $newRole = Role::create([
-            'name' => 'Default',
-            'handle' => 'default',
-            'permissions' => new Collection([1]),
-        ])->save();
-
-        $newRole->delete();
-        $role->delete();
-        $permission->delete();
+        $this->assertInstanceOf(Collection::class, $extensions);
+        $this->assertEquals(2, $extensions?->first()?->getId());
     }
 
     /**